requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
paddlepaddle==2.6.2
paddleocr==2.7.3
PyMuPDF==1.24.9
Pillow==10.2.0
python-multipart==0.0.6
opencv-python-headless==4.9.0.80
numpy==1.26.3
shapely==2.0.3
